Создает functions.php + правки
